Allow HTTPS secure protocol for MathJax CDN
The [MathJax official documentation](http://mathjax.readthedocs.org/en/latest/misc/faq.html#how-do-i-access-the-mathjax-cdn-using-a-https-secure-connection) says, "The MathJax CDN can be accessed via https. We advise to always use https."
Making the CDN url start with `//` instead of the usual `http://` allows deployment web servers to automatically switch to either http or https connection when connecting to the MathJax CDN based on the connection protocol of the hosting sites.